# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DWLite Open API",
        "description": "\u7403\u978b\u4f9b\u5e94\u94fe\u8d44\u6e90\u64ae\u5408\u4e0e\u6e20\u9053\u5206\u9500\u5e73\u53f0 Open API\n\n## \u8ba4\u8bc1\u65b9\u5f0f\n\n\u6240\u6709 API \u8bf7\u6c42\u9700\u8981\u4ee5\u4e0b Header:\n- `X-Api-Key`: API Key ID\n- `X-Timestamp`: Unix \u65f6\u95f4\u6233\uff08\u79d2\uff09\n- `X-Nonce`: \u552f\u4e00 UUID\uff08\u9632\u91cd\u653e\uff09\n- `X-Signature`: HMAC-SHA256 \u7b7e\u540d\n\n## \u7b7e\u540d\u7b97\u6cd5\n\n```\ncanonical = METHOD + \"\\n\" + PATH + \"\\n\" + TIMESTAMP + \"\\n\" + NONCE + \"\\n\" + SHA256(BODY)\nsignature = \"sha256=\" + HMAC-SHA256(canonical, API_SECRET)\n```\n\n## \u54cd\u5e94\u683c\u5f0f\n\n\u6210\u529f\u54cd\u5e94:\n```json\n{\n    \"success\": true,\n    \"data\": { ... },\n    \"requestId\": \"uuid\"\n}\n```\n\n\u9519\u8bef\u54cd\u5e94:\n```json\n{\n    \"success\": false,\n    \"error\": {\n        \"code\": \"ERROR_CODE\",\n        \"message\": \"\u9519\u8bef\u63cf\u8ff0\"\n    },\n    \"requestId\": \"uuid\"\n}\n```\n",
        "contact": {
            "name": "DWLite API Support"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1/open",
            "description": "Open API"
        }
    ],
    "paths": {
...
        "/api/v1/open/merchant/fulfillments": {
            "get": {
                "tags": [
                    "Merchant - Fulfillment"
                ],
                "summary": "List fulfillments.",
                "operationId": "get_open_api_merchant_fulfillment_list",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "allocated",
                                "accepted",
                                "rejected",
                                "picking",
                                "packing",
                                "shipped",
                                "expired"
                            ],
                            "nullable": true
                        }
                    },
                    {
                        "name": "createdFrom",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "createdTo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "nullable": true
                        }
                    },
                    {
                        "name": "fulfillmentNo",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 50,
                            "nullable": true
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/open/merchant/fulfillments/{fulfillmentNo}": {
            "get": {
                "tags": [
                    "Merchant - Fulfillment"
                ],
                "summary": "Get fulfillment details.",
                "operationId": "get_open_api_merchant_fulfillment_details",
                "parameters": [
                    {
                        "name": "fulfillmentNo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/open/merchant/fulfillments/{fulfillmentNo}/accept": {
            "post": {
                "tags": [
                    "Merchant - Fulfillment"
                ],
                "summary": "Accept fulfillment.",
                "operationId": "post_open_api_merchant_fulfillment_accept",
                "parameters": [
                    {
                        "name": "fulfillmentNo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AcceptFulfillmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/open/merchant/fulfillments/{fulfillmentNo}/reject": {
            "post": {
                "tags": [
                    "Merchant - Fulfillment"
                ],
                "summary": "Reject fulfillment.",
                "operationId": "post_open_api_merchant_fulfillment_reject",
                "parameters": [
                    {
                        "name": "fulfillmentNo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RejectFulfillmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AcceptFulfillmentRequest": {
                "required": [
                    "acceptedAt"
                ],
                "properties": {
                    "acceptedAt": {
                        "type": "string"
                    },
                    "notes": {
                        "type": "string",
                        "maxLength": 500,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "RejectFulfillmentRequest": {
                "required": [
                    "reason",
                    "rejectedAt"
                ],
                "properties": {
                    "reason": {
                        "type": "string",
                        "maxLength": 500
                    },
                    "rejectedAt": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API Key ID",
                "name": "X-Api-Key",
                "in": "header"
            },
            "TimestampAuth": {
                "type": "apiKey",
                "description": "Unix \u65f6\u95f4\u6233\uff08\u79d2\uff09",
                "name": "X-Timestamp",
                "in": "header"
            },
            "NonceAuth": {
                "type": "apiKey",
                "description": "\u552f\u4e00 UUID\uff08\u9632\u91cd\u653e\u653b\u51fb\uff09",
                "name": "X-Nonce",
                "in": "header"
            },
            "SignatureAuth": {
                "type": "apiKey",
                "description": "HMAC-SHA256 \u7b7e\u540d",
                "name": "X-Signature",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "ApiKeyAuth": [],
            "TimestampAuth": [],
            "NonceAuth": [],
            "SignatureAuth": []
        }
    ],
...
}